requirements.txt

pandas
datasets
chromadb
langgraph
openai
IPython
langchain-openai
langchain


# Core dependencies
#openai==1.52.0         
#langgraph==0.1.0       
#langchain==0.1.0       
langgraph>=0.2.0       
langchain>=0.2.0       


# Utilities
python-dotenv==1.0.0    
requests==2.31.0        


# Optional: For development and testing
pytest==7.4.0          # Testing framework
black==23.0.0          # Code formatter
pylint==2.17.0         # Code linter
